beginnings

+++ dan_layer/storage_sqlite/migrations/2022-09-13-174523_templates/up.sql
@@ -3,12 +3,13 @@ create table templates (
     -- the address is the hash of the content
     template_address    blob not null,
     -- where to find the template code
-    url                 text not null,                              
+    url                 text not null,
     -- the block height in which the template was published
-    height              bigint not null,
-    -- compiled template code as a WASM binary 
+    height              integer not null,
+    -- compiled template code as a WASM binary
     compiled_code       blob not null
 );
+
 
 -- fetching by the template_address will be a very common operation
 create index templates_template_address_index on templates (template_address);
